+++ src/test/scala/org/podval/tools/testing/ScalaOnlyTest.scala
@@ -1,31 +1,17 @@
 package org.podval.tools.testing
 
-import ForClass.*
+import org.podval.tools.build.ScalaLibrary
 
 class ScalaOnlyTest extends GroupingFunSpec:
   describe("scala-only"):
-    test(
-      project = Memo(TestProject.existingProject("scala-only")),
-      checks = Seq(
-        forClass(className = "ExampleSpec",
-          failedCount(0),
-          skippedCount(0),
-          passed("The Scala language must add correctly")
-        ),
-        forClass(className = "StackSpec",
-          failedCount(0),
-          skippedCount(1),
-          passed("A Stack should pop values in last-in-first-out order"),
-          skipped("A Stack should throw NoSuchElementException if an empty stack is popped")
-        ),
-        forClass(className = "TwoByTwoTest",
-          failedCount(1),
-          skippedCount(0),
-          passed("2*2 success should work"),
-          passed("3*3 success should work"),
-          failed("2*2 failure should fail")
-        )
-      ),
-      commandLineIncludeTestNames = Seq.empty
+    groupTest(
+      features = Seq(ScalaOnlyFeature),
+      fixtures = Seq(ScalaOnlyFixture),
+      platforms = Seq(Platform(ScalaLibrary.Scala3.versionDefault, isScalaJS = false)),
+      groupByFeature = true,
+      combinedFixtureNameOpt = None
     )
-
+    //  testLogging {
+    ////    events "started", "skipped", "failed", "passed", "standard_error", "standard_out"
+    ////    events "failed", "standard_error", "standard_out"
+    //  }