+++ test-projects/scala-only/src/test/scala/org/podval/tools/testing/Nested.scala
@@ -1,9 +1,0 @@
-package org.podval.tools.testing
-
-import org.scalatest._
-
-class Nested extends Suites(
-//  new ExampleSpec,
-//  new StackSpec,
-//  new TwoByTwoTest
-)